pyproject.toml
# View options here: https://python-poetry.org/docs/configuration
[tool.poetry]
description = ""
name = "mcq_bot"
authors = ["extrange <29305375+extrange@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.36.0"
uvloop = "^0.19.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.32"
sqlalchemy-utils = "^0.41.2"
openpyxl = "^3.1.5"
python-dateutil = "^2.9.0.post0"
schedule = "^1.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pyright = "^1.1.375"
ruff = "^0.5.7"
pytest = "^8.3.2"

[tool.poetry.scripts]
start = "mcq_bot.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"

# View rules here: https://docs.astral.sh/ruff/rules/
# [tool.ruff.lint]
# select = ["ALL"]
# ignore = [ "D211", "D212"]
